% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggUtils-package.R
\docType{package}
\name{bggUtils-package}
\alias{bggUtils}
\alias{bggUtils-package}
\title{bggUtils: Functions for BoardGameGeek Data}
\description{
Package containing helper functions for loading and analysing BGG data. Makes use of bggAnalytics for data collection and general tidyverse functions for manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phenrickson/bggUtils}
  \item Report bugs at \url{https://github.com/phenrickson/bggUtils/issues}
}

}
\keyword{internal}
