% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bgg_text.R
\name{clean_bgg_text}
\alias{clean_bgg_text}
\title{clean_bgg_text: standardize bgg text variables for use in modeling}
\usage{
clean_bgg_text(text)
}
\arguments{
\item{text}{text to be cleaned}
}
\description{
takes bgg categorical information and applies standardized preprocessing for creating features for modeling
}
\examples{
clean_bgg_text("Fantasy Flight Games")
}
