% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{clean_text: standardize text variables for use in modeling}
\usage{
clean_text(text)
}
\arguments{
\item{text}{text to be cleaned}
}
\description{
takes bgg categorical information and applies standardized preprocessing for creating features for modeling
}
\examples{
clean_text("Fantasy Flight Games")
}
