% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_playercounts.R
\name{find_playercounts}
\alias{find_playercounts}
\title{filter wide player count data to find games matching crtieria}
\usage{
find_playercounts(data, type = c("best"), numplayers)
}
\arguments{
\item{data}{wide player count data}

\item{type}{type of vote, such as 'best', 'recommended', or 'not recommended'}

\item{numplayers}{number of players}
}
\value{
data frame with game ids matching type and numplayers
}
\description{
filter wide player count data to find games matching crtieria
}
\examples{
query_playercounts(game_id = c(12, 7, 21)) |>
  wider_playercounts() |>
  find_playercounts(type = "best", numplayers = 3)

}
