% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bgg_games.R
\name{get_bgg_games}
\alias{get_bgg_games}
\title{get_bgg_games: pulling data for games on BoardGameGeek}
\usage{
get_bgg_games(game_ids, batch_size = 500, tidy = T, toJSON = F)
}
\arguments{
\item{game_ids}{vector of game_ids}

\item{batch_size}{number of game_ids to submit to BGG in one batch; defaults to maximum value of 500}

\item{tidy}{return tidied dataframe or raw xml; defaults to T}

\item{toJSON}{convert results to json; defaults to F}
}
\description{
This function submits game ids to BGG's API and returns data associated with a game.
}
\examples{
get_bgg_games(c(12, 7))
}
