% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bgg_ids.R
\name{get_bgg_ids}
\alias{get_bgg_ids}
\title{get_bgg_ids: load universe of ids for games and expansions on boardgamegeek}
\usage{
get_bgg_ids(id_type = NULL, names = c("id", "type"))
}
\arguments{
\item{id_type}{boardgame, boardgameexpansion, boardgameaccessory}

\item{names}{character vector of length two with names for columns}
}
\description{
This function loads the universe of BGG ids from https://bgg.activityclub.org/bggdata/thingids.txt
}
