% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games_from_gcp.R
\name{get_games_from_gcp}
\alias{get_games_from_gcp}
\title{load games data from gcp bucket}
\usage{
get_games_from_gcp(
  bucket_name = "bgg_data",
  object = "raw/objects/games",
  generation = NULL
)
}
\arguments{
\item{bucket_name}{Name of the GCS bucket}

\item{object}{Name of the object to load}

\item{generation}{Optional generation to load}
}
\value{
Data frame of games or NULL if not found
}
\description{
load games data from gcp bucket
}
