% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{get query via DBI}
\usage{
get_query(query, conn, eval = T)
}
\arguments{
\item{query}{a SQL query as a string}

\item{conn}{a DBI connection}

\item{eval}{to evaluate the query or to just show the query with connection}
}
\description{
runs query given a connection
}
