% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_collection.R
\name{get_user_collection}
\alias{get_user_collection}
\title{get_user_collection: pulling collection data for a boardgamegeek user}
\usage{
get_user_collection(username)
}
\arguments{
\item{username}{a username on boardgamegeek; case sensitive}
}
\value{
data frame with user collection
}
\description{
This function pulls boardgamegeek collection data for a given username
}
\examples{
user_collection <- get_user_collection("mrbananagrabber")


}
