% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_bgg_games.R
\name{preprocess_bgg_games}
\alias{preprocess_bgg_games}
\title{preprocess_bgg_games: standardized preprocessing for bgg api requests for use in modeling}
\usage{
preprocess_bgg_games(
  data,
  publisher_allow = publishers_allow_list(),
  families_allow = families_allow_list(),
  families_remove = families_remove_list()
)
}
\arguments{
\item{data}{data from `get_bgg_games()` with simplify=T}

\item{publisher_allow}{list of allowed publisher ids in modeling}

\item{families_allow}{allow list of permitted bgg family variables for use in modeling}

\item{families_remove}{list of bgg family variables to be removed for modeling}
}
\description{
takes output of `get_bgg_games()` and tidies output for modeling. this includes replacing 0s with NAs for numeric values,
filtering to include/exclude selected categorical features, and collapsing categorical features for use with tokenization
}
\examples{
api_resp <- get_bgg_games(c(12, 7), simplify = TRUE)

api_resp \%>\%
  unnest(data) \%>\%
  preprocess_bgg_games()

}
