% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_bgg_text.R
\name{present_bgg_text}
\alias{present_bgg_text}
\title{present_bgg_text: clean bgg text/variables for presentation in plots, tables}
\usage{
present_bgg_text(text, minlength = 50)
}
\arguments{
\item{text}{text to be cleaned}

\item{minlength}{minimum length of abbreviated ext returned for presentation}
}
\description{
takes bgg string and applies standardized preprocessing for presentation
}
\examples{
present_bgg_text("fantasy_flight_games")
}
