% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_text.R
\name{present_text}
\alias{present_text}
\title{present_text: clean bgg variables for presentation in plots, tables}
\usage{
present_text(text, minlength = 30)
}
\arguments{
\item{text}{text to be cleaned}

\item{minlength}{minimum length of abbreviated ext returned for presentation}
}
\description{
takes bgg string and applies standardized preprocessing for presentation
}
\examples{
clean_text("Fantasy Flight Games")
}
