% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_games.R
\name{query_games}
\alias{query_games}
\title{query active games data}
\usage{
query_games(conn = bigquery_connect(), game_ids = NULL, eval = T)
}
\arguments{
\item{conn}{DBI connection to database}

\item{game_ids}{vector of game ids}

\item{eval}{whether to evaluate query or not}
}
\value{
data frame with info for selected game ids
}
\description{
query active games data
}
