% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_info.R
\name{query_info}
\alias{query_info}
\title{query info about selected game ids}
\usage{
query_info(conn = bigquery_connect(), game_ids, eval = T)
}
\arguments{
\item{conn}{DBI connection to database}

\item{game_ids}{vector of game ids}

\item{eval}{whether to evaluate query or not}
}
\value{
data frame with info for selected game ids
}
\description{
query info about selected game ids
}
