% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_playercounts.R
\name{query_playercounts}
\alias{query_playercounts}
\title{query playercount data for selected games}
\usage{
query_playercounts(
  conn = bigquery_connect(),
  game_ids,
  rec_values = c("Best", "Recommended", "Not Recommended"),
  vote_types = c("yes"),
  minvotes = 10,
  eval = T,
  ...
)
}
\arguments{
\item{conn}{DBI connection to database}

\item{game_ids}{vector of game ids}

\item{rec_values}{type of recommendation ('Best', 'Not Recommended')}

\item{vote_types}{vote types}

\item{minvotes}{minimum number of votes}

\item{eval}{whether to evaluate query or not}

\item{...}{}
}
\value{
a data frame with number of votes for playercounts
}
\description{
query playercount data for selected games
}
