% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_games.R
\name{request_games}
\alias{request_games}
\title{request game data from BGG API (v2)}
\usage{
request_games(game_ids, ...)
}
\arguments{
\item{game_ids}{vector of game_ids}

\item{...}{additional arguments to pass to request request bgg api}
}
\description{
This function submits game ids to BGG's API and returns data associated with a game in a nested table
}
\details{
#' @description
}
\examples{
request_games(game_ids = c(12, 7))
}
