% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_deserialize.R
\name{try_deserialize}
\alias{try_deserialize}
\title{Try to deserialize data using either qs or qs2 package}
\usage{
try_deserialize(data)
}
\arguments{
\item{data}{Raw data to deserialize}
}
\value{
Deserialized data
}
\description{
Try to deserialize data using either qs or qs2 package
}
\keyword{internal}
