% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wider_playercounts.R
\name{wider_playercounts}
\alias{wider_playercounts}
\title{pivot and summarize longer player count data for presentation in tables}
\usage{
wider_playercounts(data)
}
\arguments{
\item{data}{plong layercounts data by game}
}
\value{
playercounts data in wide format
}
\description{
pivot and summarize longer player count data for presentation in tables
}
