% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_connect.R
\name{bigquery_connect}
\alias{bigquery_connect}
\title{connect to bigquery}
\usage{
bigquery_connect(project = Sys.getenv("GCS_PROJECT_ID"), dataset = "bgg")
}
\arguments{
\item{project}{gcp project id}

\item{dataset}{bigquery schema}
}
\description{
connect to bigquery
}
