% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games_from_gcp.R
\name{get_games_from_gcp}
\alias{get_games_from_gcp}
\title{load games data from gcp bucket}
\usage{
get_games_from_gcp(
  bucket = "bgg_data",
  object_name = games_object(),
  generation = NULL,
  ...
)
}
\arguments{
\item{bucket}{gcp bucket}

\item{object_name}{path to object in gcp bucket}

\item{generation}{version of object in bucket}
}
\value{
games data in nested data frame
}
\description{
load games data from gcp bucket
}
